http client shutdown in example
